+++ Source/EventFlow.TestHelpers/Aggregates/Sagas/ThingySagaErrorHandler.cs
@@ -1,4 +1,4 @@
-// The MIT License (MIT)
+﻿// The MIT License (MIT)
 // 
 // Copyright (c) 2015-2020 Rasmus Mikkelsen
 // Copyright (c) 2015-2020 eBay Software Foundation
@@ -21,22 +21,19 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using EventFlow.Sagas;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace EventFlow.Sagas
+namespace EventFlow.TestHelpers.Aggregates.Sagas
 {
-    public class SagaErrorHandler : ISagaErrorHandler
+    public class ThingySagaErrorHandler : ISagaErrorHandler<ThingySaga>
     {
-        public Task<bool> HandleAsync(
-            ISagaId sagaId,
-            SagaDetails sagaDetails,
-            Exception exception,
+        public Task<bool> HandleAsync(ISagaId sagaId, SagaDetails sagaDetails, Exception exception,
             CancellationToken cancellationToken)
         {
-            // The default handler cannot handle anything!
-            return Task.FromResult(false);
+            return Task.FromResult(true);
         }
     }
 }