+++ Source/EventFlow/Snapshots/ISnapshotMetadata.cs
@@ -21,6 +21,7 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System.Collections.Generic;
 using EventFlow.Core;
 
 namespace EventFlow.Snapshots
@@ -32,5 +33,7 @@ namespace EventFlow.Snapshots
         int AggregateSequenceNumber { get; }
         string AggregateId { get; }
         string AggregateName { get; }
+        IReadOnlyCollection<ISourceId> PreviousSourceIds { get;  }
+
     }
 }