+++ RELEASE_NOTES.md
@@ -1,4 +1,15 @@
-### New in 0.80 (not released yet)
+### New in 0.82 (not released yet)
+
+* Fix: Source IDs are now added to snapshots
+
+### New in 0.81.4483 (released 2020-12-14)
+
+* Breaking: Elasticsearch NEST Nuget Library updated from v6.1.0 to v7.8.2
+* New: Now possible to implement error handlers for specific sagas using
+  `ISagaErrorHandler<TSaga>`
+* Fixed: You can now create `Id : Identity<Id>`
+
+### New in 0.80.4377 (released 2020-10-01)
 
 * Breaking: To support .NET going forward, all EventFlow test have been converted
   from .NET Framework 4.x to .NET Core 3.1. This however, introduced a set of
@@ -28,7 +39,7 @@
   * `IDispatchToSubscriberResilienceStrategy`
   * `ISagaUpdateResilienceStrategy`
 
-### New in 0.79.4216 ((released 2020-05-13)
+### New in 0.79.4216 (released 2020-05-13)
 
 * New: Added .NET Core 3.1 target for the `EventFlow`
   and `EventFlow.EntityFramework` packages